Delete TypeDefault call code generation logic in VariableType
There is a new invariant that emit_body is only ever called when
strategy is 'use_derived', which means we can delete a bunch of code.
This removes the last use of TypeXXX.h headers.

Note that this change makes sense, as the TypeDefault entries are
registered as Math entries, which means they automatically populate
Autograd (and we no longer have to register them ourselves).  Ailing
did all the hard work, this is just the payoff.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 95da6a366a9969228d3e6ae5dae27b2c5e0de972
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47000