[NNAPI] Remove solid weights support
Summary:
This was left over from an old version of the code.  The idea was that
instead of indexing into separate tensors for each weight, you could
bundle them all into a single file and use different offsets into that
file.  With the current design, this is nontrivial to support, so drop
the code for now.

Test Plan:
CI

ghstack-source-id: 9156e5f87fecbaf85d46256193a8accb432c779f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47518